apps/backend/migrations/versions/001_criar_tabelas.py
"""Criar tabelas courses, lessons e questions

Revision ID: 001
Revises: 
Create Date: 2026-03-07
"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa
from sqlalchemy.dialects.postgresql import JSONB

revision: str = '001'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    op.create_table(
        'courses',
        sa.Column('id', sa.Integer, primary_key=True),
        sa.Column('title', sa.String, nullable=False),
        sa.Column('description', sa.String),
        sa.Column('icon', sa.String),
        sa.Column('color', sa.String),
        sa.Column('total_lessons', sa.Integer),
    )

    op.create_table(
        'lessons',
        sa.Column('id', sa.Integer, primary_key=True),
        sa.Column('course_id', sa.Integer,
                  sa.ForeignKey('courses.id', ondelete='CASCADE'), nullable=False),
        sa.Column('title', sa.String, nullable=False),
        sa.Column('description', sa.String),
        sa.Column('order', sa.Integer),
    )

    op.create_table(
        'questions',
        sa.Column('id', sa.Integer, primary_key=True),
        sa.Column('lesson_id', sa.Integer,
                  sa.ForeignKey('lessons.id', ondelete='CASCADE'), nullable=False),
        sa.Column('question', sa.String, nullable=False),
        sa.Column('code', sa.String),
        sa.Column('options', JSONB),
        sa.Column('correct_answer', sa.Integer),
        sa.Column('order', sa.Integer),
    )


def downgrade() -> None:
    op.drop_table('questions')
    op.drop_table('lessons')
    op.drop_table('courses')