apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4", # Migrations
    "flasgger>=0.9.7.1", # Swagger
    "flask>=3.1.3", # Framework
    "flask-sqlalchemy>=3.1.1", # ORM
    "psycopg2-binary>=2.9.11", # Driver Postgres
    "pydantic>=2.12.5", # Validação
    "python-dotenv>=1.2.1",
]
